Moved RadioConfidenceConverter into DataHelper.cs
