+++ VSProjects/Geomapmaker/DataHelper.cs
@@ -9,9 +9,11 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Data;
 using System.Windows.Media;
 
 namespace Geomapmaker
@@ -927,6 +929,22 @@ namespace Geomapmaker
             {
                 return locationConfidenceMeters;
             }
+        }
+    }
+
+    /// <summary>
+    /// Value converter for radio button groups
+    /// </summary>
+    public class RadioConfidenceConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return Equals(value, parameter);
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return parameter;
         }
     }
 }
