+++ VSProjects/Geomapmaker/Views/AddEditContactsAndFaultsDockPane.xaml
@@ -72,7 +72,7 @@
 
         <ToggleButton Grid.Row="2" Grid.Column="1" Grid.ColumnSpan="3" x:Name="PopulateFromExistingToggle" Content="Populate values from existing line" IsChecked="{Binding Prepopulate, Mode=TwoWay}" HorizontalAlignment="Center" Padding="5" Margin="5"/>
 
-        <TextBlock Grid.Row="3" Grid.Column="3" Margin="5" Text="Current Values" VerticalAlignment="Center" TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}" />
+        <TextBlock Grid.Row="3" Grid.Column="3" Margin="5" FontWeight="Bold" Text="Current Values" VerticalAlignment="Center" TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}" />
 
         <TextBlock Grid.Row="4" Grid.Column="1" Margin="5" Text="Type:" VerticalAlignment="Center" TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}" />
         <ComboBox
@@ -154,7 +154,7 @@
               >
             </RadioButton>
         </Grid>
-        <TextBlock Grid.Row="5" Grid.Column="3" Margin="5" Text="{Binding SelectedCF.IdentityConfidence}" ToolTip="Identity Confidence" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
+        <TextBlock Grid.Row="5" Grid.Column="3" Margin="5" FontWeight="Bold" Text="{Binding SelectedCF.IdentityConfidence}" ToolTip="Identity Confidence" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
 
         <TextBlock Grid.Row="6" Grid.Column="1" Margin="5" Text="Existence Confidence:" TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}" />
         <Grid Grid.Row="6" Grid.Column="2" >
@@ -212,7 +212,7 @@
               >
             </RadioButton>
         </Grid>
-        <TextBlock Grid.Row="6" Grid.Column="3" Margin="5" Text="{Binding SelectedCF.ExistenceConfidence}" ToolTip="Existence Confidence" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
+        <TextBlock Grid.Row="6" Grid.Column="3" Margin="5" FontWeight="Bold" Text="{Binding SelectedCF.ExistenceConfidence}" ToolTip="Existence Confidence" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
 
         <TextBlock Grid.Row="7" Grid.Column="1" Margin="5" Text="Location Confidence Meters:" TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}" />
         <ComboBox 
@@ -221,21 +221,21 @@
             x:Name="LocationConfidenceComboBox" 
             Margin="5" 
             ItemsSource="{Binding Path=(ui:DataHelper.LocationConfidenceMeters)}" 
-            SelectedValue="{Binding SelectedCF.LocationConfidenceMeters, Mode=TwoWay}"
+            SelectedValue="{Binding SelectedCF.LocationConfidenceMeters}"
             IsReadOnly="False" 
             IsEditable="True" 
             HorizontalAlignment="Stretch" 
             SelectedValuePath="Key"
             DisplayMemberPath="Value" 
             />
-        <TextBlock Grid.Row="7" Grid.Column="3" Margin="5" VerticalAlignment="Center" Text="{Binding SelectedCF.LocationConfidenceMeters}" ToolTip="Location Confidence" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
+        <TextBlock Grid.Row="7" Grid.Column="3" Margin="5" FontWeight="Bold" VerticalAlignment="Center" Text="{Binding SelectedCF.LocationConfidenceMeters}" ToolTip="Location Confidence" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
 
         <TextBlock Grid.Row="8" Grid.Column="1" Margin="5" Text="Concealed:" Style="{DynamicResource Esri_TextBlockRegular}" />
-        <CheckBox Grid.Row="8" Grid.Column="2" Style="{DynamicResource Esri_CheckboxToggleSwitch}" HorizontalAlignment="Right" IsChecked="{Binding SelectedCF.IsConcealed}" ></CheckBox>
-        <TextBlock Grid.Row="8" Grid.Column="3" Margin="5" Text="{Binding SelectedCF.IsConcealed}" ToolTip="Concealed" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
+        <CheckBox Grid.Row="8" Grid.Column="2" Style="{DynamicResource Esri_CheckboxToggleSwitch}" HorizontalAlignment="Right" IsChecked="{Binding SelectedCF.IsConcealed}" Margin="5" ></CheckBox>
+        <TextBlock Grid.Row="8" Grid.Column="3" Margin="5" FontWeight="Bold" Text="{Binding SelectedCF.IsConcealed}" ToolTip="Concealed" Style="{DynamicResource Esri_TextBlockRegular}" HorizontalAlignment="Center" />
 
         <TextBlock Grid.Row="9" Grid.Column="1" Margin="5" Text="Notes:" TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}"/>
-        <TextBox Grid.Row="9" Grid.Column="2" x:Name="NotesTextBlock" Margin="5" Text="{Binding SelectedCF.Notes}" HorizontalAlignment="Stretch" TextWrapping="Wrap" />
+        <TextBox Grid.Row="9" Grid.Column="2" x:Name="NotesTextBlock" Margin="5" Text="{Binding SelectedCF.Notes}" HorizontalAlignment="Stretch" TextWrapping="Wrap" MaxLength="250" />
 
         <WrapPanel Grid.Row="10" Grid.Column="1" Margin="5">
             <TextBlock Text="Data Source: " TextWrapping="Wrap" Style="{DynamicResource Esri_TextBlockRegular}"/>
